+++ src/walletFeature/wallet/modals/SelectMethodModalContent/types.ts
@@ -1,6 +1,7 @@
 import { AppIconName } from '@app/walletFeature/wallet/common/components/AppIcon/types';
 
 export type SelectMethodModalItem = {
+  disabled?: boolean;
   icon: AppIconName;
   title: string;
   subtitle: string;