+++ src/walletFeature/wallet/screens/WalletDetails/WalletDetailsScreen.tsx
@@ -8,10 +8,7 @@ import { RouteProp, useRoute } from '@react-navigation/native';
 import { WalletParamList } from '@app/walletFeature/wallet/navigation/types';
 import { WalletRoute } from '@app/walletFeature/wallet/navigation/constants';
 import { useAppTheme } from '@app/walletFeature/wallet/common/theme';
-import {
-  formatNumber,
-  getDecimals,
-} from '@app/walletFeature/wallet/common/utils/number';
+import { formatNumber } from '@app/walletFeature/wallet/common/utils/number';
 import { AppImage } from '@app/walletFeature/wallet/common/components/AppImage/AppImage';
 import { NetworkItem } from '@app/walletFeature/wallet/screens/WalletDetails/components/NetworkItem/NetworkItem';
 import { HistoryTabContent } from '@app/walletFeature/wallet/components/HistoryTabContent/HistoryTabContent';
@@ -44,11 +41,11 @@ export const WalletDetailsScreen: FC = () => {
           </AppView>
         )}
 
-        <AppText textStyle="medium_26_32">{`${formatNumber(
+        <AppText textAlign="center" textStyle="medium_26_32">{`${formatNumber(
           Number(item?.totalBalanceAcrossNetworks.balance ?? 0),
           undefined,
           2,
-          getDecimals(item.networks[0]?.asset.decimals),
+          item.networks[0].asset.decimals,
         )} ${item.symbol}`}</AppText>
         <AppText
           color={colors.inputLabelColor}
@@ -56,7 +53,7 @@ export const WalletDetailsScreen: FC = () => {
           Number(item?.totalBalanceAcrossNetworks.balanceUsd ?? 0),
           'currency',
           2,
-          getDecimals(item.networks[0]?.asset.decimals),
+          item.networks[0]?.asset.decimals,
         )}`}</AppText>
       </AppView>
       <AppView